Python/requirements.txt
gpiozero
numpy
paho-mqtt
python-crontab
pyyaml
smbus
